+++ huaweicloud-sdk-iotda/huaweicloudsdkiotda/v5/region/iotda_region.py
@@ -30,6 +30,8 @@ class IoTDARegion:
                         "https://iotda.sa-brazil-1.myhuaweicloud.com")
     LA_NORTH_2 = Region("la-north-2",
                         "https://iotda.la-north-2.myhuaweicloud.com")
+    RU_MOSCOW_1 = Region("ru-moscow-1",
+                        "https://iotda.ru-moscow-1.myhuaweicloud.com")
 
     static_fields = {
         "cn-north-4": CN_NORTH_4,
@@ -44,6 +46,7 @@ class IoTDARegion:
         "la-south-2": LA_SOUTH_2,
         "sa-brazil-1": SA_BRAZIL_1,
         "la-north-2": LA_NORTH_2,
+        "ru-moscow-1": RU_MOSCOW_1,
     }
 
     @classmethod