+++ huaweicloud-sdk-oms/huaweicloudsdkoms/v2/model/task_resp.py
@@ -52,6 +52,7 @@ class TaskResp:
         'dst_storage_policy': 'str',
         'consistency_check': 'str',
         'enable_requester_pays': 'bool',
+        'obs_system': 'str',
         'task_priority': 'str'
     }
 
@@ -93,10 +94,11 @@ class TaskResp:
         'dst_storage_policy': 'dst_storage_policy',
         'consistency_check': 'consistency_check',
         'enable_requester_pays': 'enable_requester_pays',
+        'obs_system': 'obs_system',
         'task_priority': 'task_priority'
     }
 
-    def __init__(self, bandwidth_policy=None, complete_size=None, description=None, dst_node=None, enable_failed_object_recording=None, enable_metadata_migration=None, enable_restore=None, error_reason=None, failed_num=None, failed_object_record=None, group_id=None, id=None, is_query_over=None, left_time=None, migrate_since=None, migrate_speed=None, name=None, progress=None, real_size=None, skipped_num=None, src_node=None, start_time=None, status=None, successful_num=None, task_type=None, group_type=None, total_num=None, total_size=None, total_time=None, smn_info=None, source_cdn=None, success_record_error_reason=None, skip_record_error_reason=None, object_overwrite_mode=None, dst_storage_policy=None, consistency_check=None, enable_requester_pays=None, task_priority=None):
+    def __init__(self, bandwidth_policy=None, complete_size=None, description=None, dst_node=None, enable_failed_object_recording=None, enable_metadata_migration=None, enable_restore=None, error_reason=None, failed_num=None, failed_object_record=None, group_id=None, id=None, is_query_over=None, left_time=None, migrate_since=None, migrate_speed=None, name=None, progress=None, real_size=None, skipped_num=None, src_node=None, start_time=None, status=None, successful_num=None, task_type=None, group_type=None, total_num=None, total_size=None, total_time=None, smn_info=None, source_cdn=None, success_record_error_reason=None, skip_record_error_reason=None, object_overwrite_mode=None, dst_storage_policy=None, consistency_check=None, enable_requester_pays=None, obs_system=None, task_priority=None):
         r"""TaskResp
 
         The model defined in huaweicloud sdk
@@ -175,6 +177,8 @@ class TaskResp:
         :type consistency_check: str
         :param enable_requester_pays: 是否开启请求者付款，在启用后，请求者支付请求和数据传输费用。
         :type enable_requester_pays: bool
+        :param obs_system: OBS系统类型 BUCKET：一般桶 PFS：并行文件系统
+        :type obs_system: str
         :param task_priority: HIGH：高优先级 MEDIUM：中优先级 LOW：低优先级
         :type task_priority: str
         """
@@ -218,6 +222,7 @@ class TaskResp:
         self._dst_storage_policy = None
         self._consistency_check = None
         self._enable_requester_pays = None
+        self._obs_system = None
         self._task_priority = None
         self.discriminator = None
 
@@ -295,6 +300,8 @@ class TaskResp:
             self.consistency_check = consistency_check
         if enable_requester_pays is not None:
             self.enable_requester_pays = enable_requester_pays
+        if obs_system is not None:
+            self.obs_system = obs_system
         if task_priority is not None:
             self.task_priority = task_priority
 
@@ -1087,6 +1094,28 @@ class TaskResp:
         :type enable_requester_pays: bool
         """
         self._enable_requester_pays = enable_requester_pays
+
+    @property
+    def obs_system(self):
+        r"""Gets the obs_system of this TaskResp.
+
+        OBS系统类型 BUCKET：一般桶 PFS：并行文件系统
+
+        :return: The obs_system of this TaskResp.
+        :rtype: str
+        """
+        return self._obs_system
+
+    @obs_system.setter
+    def obs_system(self, obs_system):
+        r"""Sets the obs_system of this TaskResp.
+
+        OBS系统类型 BUCKET：一般桶 PFS：并行文件系统
+
+        :param obs_system: The obs_system of this TaskResp.
+        :type obs_system: str
+        """
+        self._obs_system = obs_system
 
     @property
     def task_priority(self):